refactor: moved everything from /model/commands to /commands
