+++ src/main/java/com/seailz/discordjv/DiscordJv.java
@@ -10,9 +10,9 @@ import com.seailz.discordjv.gateway.GatewayFactory;
 import com.seailz.discordjv.model.application.Application;
 import com.seailz.discordjv.model.application.Intent;
 import com.seailz.discordjv.model.channel.Channel;
-import com.seailz.discordjv.model.commands.Command;
-import com.seailz.discordjv.model.commands.CommandChoice;
-import com.seailz.discordjv.model.commands.CommandOption;
+import com.seailz.discordjv.command.Command;
+import com.seailz.discordjv.command.CommandChoice;
+import com.seailz.discordjv.command.CommandOption;
 import com.seailz.discordjv.model.emoji.sticker.Sticker;
 import com.seailz.discordjv.model.emoji.sticker.StickerPack;
 import com.seailz.discordjv.model.guild.Guild;
