+++ src/main/java/com/seailz/discordjv/command/listeners/slash/SubCommandListener.java
@@ -3,7 +3,7 @@ package com.seailz.discordjv.command.listeners.slash;
 import com.seailz.discordjv.command.listeners.CommandListener;
 import com.seailz.discordjv.events.model.interaction.command.CommandInteractionEvent;
 import com.seailz.discordjv.events.model.interaction.command.SlashCommandInteractionEvent;
-import com.seailz.discordjv.model.commands.CommandType;
+import com.seailz.discordjv.command.CommandType;
 
 public interface SubCommandListener extends CommandListener {
