+++ src/main/java/com/seailz/discordjv/command/permissions/ApplicationCommandPermissions.java
@@ -1,4 +1,4 @@
-package com.seailz.discordjv.model.commands.permissions;
+package com.seailz.discordjv.command.permissions;
 
 import com.seailz.discordjv.core.Compilerable;
 import org.jetbrains.annotations.NotNull;
